requirements.txt
tensorflow-gpu==1.4.0
sklearn==0.0
scipy==1.1.0
